WIP on fit-all command
The current scale calculation mismatches units, working with both pixels and physical units represented by the polygon coordinates. We need to find a way to access the current size of the canvas (not window) in pixels. This can be accessed from the ScrollChangedEventArgs under ExtentHeight and ExtentWidth, but currently the OnFitAll command doesn't receive them.