Improved "Fit" button
The reason the fit all command wasn't working was that the scaling was initially being done with mixed units: dxf/svg units and pixels. Additionally, the way the polygons are rendered is that they are positioned on a canvas, and then that canvas is scaled to fit onto a fixed-size grid. This means that scaling can be done based on the window dimensions and the grid dimensions (both in pixels), and the polygon extrema (which are not in pixels) don't matter.